+++ changelog.md
@@ -5,6 +5,8 @@
 - Fix converter for older versions
 - Added a way to export/import data via `/zhomes export` and `/zhomes import <data>`
 - Fix support for Geyser-Standalone
+- Fix events loading more than one time in `/zhomes reload`
+- Added a new section to the config.yml to restrict homes in certain worlds: `restricted-worlds`
 
 
 **Full Changelog**: https://github.com/yL3oft/zHomes/compare/2.2.0...2.2.1